docker/requirements.txt
SQLAlchemy==2.0.22
pydantic-settings==2.0.3
fastapi==0.104.1
python-jose==3.3.0
SpeechRecognition==3.10.0
PyMySQL==1.1.0
uvicorn==0.23.2
python-multipart==0.0.6